+++ SequelNet.Connector.MySql/MySqlLanguageFactory.cs
@@ -482,7 +482,7 @@ namespace SequelNet.Connector
             else if (dataType == DataType.UnsignedSmallInt)
                 sb.Append("SMALLINT UNSIGNED");
             else if (dataType == DataType.Int)
-                sb.Append("TINYINT");
+                sb.Append("INT");
             else if (dataType == DataType.UnsignedInt)
                 sb.Append("INT UNSIGNED");
             else if (dataType == DataType.BigInt)
